feat: validation de psicologos funcionando

+++ src/controller/psicologoController.js
@@ -1,20 +1,19 @@
 const { Psicologos, Atendimentos } = require("../models");
-const bcrypt = require('bcryptjs');
+const bcrypt = require("bcryptjs");
 
 const psicologoController = {
   async listarPsicologos(req, res) {
     const listaDePsicologos = await Psicologos.findAll({
       include: Atendimentos,
     });
-    res.json(listaDePsicologos);
+    res.status(200).json(listaDePsicologos);
   },
+
   async pegarPsicologo(req, res) {
     const { id } = req.params;
-    const psicologo = await Psicologos.findOne({
-      where: id,
-    });
+    const psicologo = await Psicologos.findByPk(id);
     if (!psicologo) {
-      res.status(404).json("ID não encontrado");
+      res.status(404).json("ID do psicologo não encontrado");
     } else {
       res.status(200).json(psicologo);
     }
@@ -29,32 +28,40 @@ const psicologoController = {
       senha: hashSenha,
       apresentacao,
     });
-    
+
     res.json(novoPsicologo);
   },
 
   async deletarPsicologo(req, res) {
     const { id } = req.params;
-
-    await Psicologos.destroy({
-      where: { id },
-    });
-
-    res.json("Successfully deleted!");
+    try {
+      const psicologoASerDeletado = await Psicologos.findByPk(id);
+      if (!psicologoASerDeletado) throw new Error("Psicologo não encontrado!");
+      await Psicologos.destroy({
+        where: { id },
+      });
+      res.status(204).json("Psicologo deletado com sucesso!");
+    }catch (error) {
+      res.status(400).json(error.message);
+    }
   },
 
   async atualizarPsicologo(req, res) {
     const { id } = req.params;
     const { nome, email, senha, apresentacao } = req.body;
-    // const newSenha = bcrypt.hashSync(senha, 10);
-    const psicologoAtualizado = await Psicologos.findOne({ where: { id } });
-    psicologoAtualizado.update({
-      nome,
-      email,
-      senha,
-      apresentacao,
-    });
-    res.status(200).json(psicologoAtualizado);
+    try {
+      const psicologoAtualizado = await Psicologos.findByPk(id);
+      if (!psicologoAtualizado) throw new Error("Psicologo não encontrado!");
+      psicologoAtualizado.update({
+        nome,
+        email,
+        senha,
+        apresentacao,
+      });
+      res.status(200).json(psicologoAtualizado);
+    } catch (error) {
+      res.status(400).json(error.message);
+    }
   },
 };
